fix: resolve material-icons KMP dependency and fix platform ImagePicker errors
- Replace AndroidX material-icons (JVM-only) with Compose Multiplatform's
  compose.materialIconsExtended for cross-platform resolution
- Fix iOS NSData.toByteArray() using memcpy instead of unresolvable pointer indexing
- Fix JS ImagePicker unresolved return label by restructuring control flow
- Fix activeDaysAgo parameter type from Long to Int to match call site

+++ composeApp/src/commonMain/kotlin/com/fugisawa/quemfaz/ui/strings/Strings.kt
@@ -178,7 +178,7 @@ object Strings {
         const val ACTIVE_RECENTLY = "Ativo(a) recentemente"
         const val COMPLETE_PROFILE = "Completar perfil"
 
-        fun activeDaysAgo(days: Long) = "Ativo(a) há $days dias"
+        fun activeDaysAgo(days: Int) = "Ativo(a) há $days dias"
     }
 
     object Errors {
